Created a specific file for paraview based functions
